helios-rtb-engine/services/analytics-service/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
psycopg2-binary==2.9.9
kafka-python==2.0.2
six>=1.16.0
python-dotenv==1.0.1
django-environ==0.11.2
gunicorn==22.0.0
